pyproject.toml
[project]
name = "flask-postgresql-gui-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "psycopg2-binary>=2.9.9",
]
